third_party/WebKit/LayoutTests/platform/oilpan/fast/dom/StyleSheet/gc-parent-rule-expected.txt

CONSOLE WARNING: 'DOMImplementation.createCSSStyleSheet()' is deprecated.
Test that CSS rules don't lose custom properties on GC when only reachable through child rules.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL rule.parentRule.foo should be bar (of type string). Was undefined (of type undefined).
PASS successfullyParsed is true

TEST COMPLETE
